+++ lib/src/components/form/fields/input.dart
@@ -6,7 +6,6 @@ import 'package:flutter/services.dart';
 import 'package:shadcn_ui/src/components/form/field.dart';
 import 'package:shadcn_ui/src/components/input.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
-import 'package:shadcn_ui/src/theme/theme.dart';
 
 class ShadInputFormField extends ShadFormBuilderField<String> {
   ShadInputFormField({
@@ -95,7 +94,6 @@ class ShadInputFormField extends ShadFormBuilderField<String> {
           onChanged: onChanged == null ? null : (v) => onChanged(v ?? ''),
           builder: (field) {
             final state = field as _ShadFormBuilderInputState;
-            final theme = ShadTheme.of(state.context);
             final effectiveDecoration = (decoration ?? const ShadDecoration())
                 .copyWith(hasError: state.hasError);
             return ShadInput(